warning fixes for plots directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23664 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/Histogram/avtHistogramFilter.C
@@ -283,7 +283,7 @@ avtHistogramFilter::PostExecute(void)
         {
             ranges = 
                  GetInput()->GetInfo().GetAttributes().GetVariableBinRanges();
-            if (ranges.size() == workingNumBins+1)
+            if (ranges.size() == (size_t)workingNumBins+1)
                 spaceBins = true;
         }
 
@@ -1030,7 +1030,7 @@ avtHistogramFilter::ArrayVarExecute(vtkDataSet *inDS, int chunk)
     int ncells = inDS->GetNumberOfCells();
     for (int i = 0 ; i < ncells ; i++)
     {
-        if (ptr[ncomps*i + compOffset] == zone)
+        if (ptr[ncomps*i + compOffset] == (size_t)zone)
         {
             workingNumBins = arr->GetNumberOfComponents();
             double *vals = new double[workingNumBins];
