+++ src/plots/Spreadsheet/QvisSpreadsheetPlotWindow.C
@@ -429,7 +429,7 @@ QvisSpreadsheetPlotWindow::UpdateSubsetNames()
             // sets that are used.
             avtSILSet_p current = restriction->GetSILSet(silTopSet);
             const std::vector<int> &mapsOut = current->GetMapsOut();
-            for(int j = 0; j < mapsOut.size(); ++j)
+            for(size_t j = 0; j < mapsOut.size(); ++j)
             {
                 int cIndex = mapsOut[j];
                 avtSILCollection_p collection =restriction->GetSILCollection(cIndex);
